Moved services to a separate file
